Fix build

+++ src/apngasm_python.cpp
@@ -237,7 +237,7 @@ NB_MODULE(MODULE_NAME, m) {
                             paletteView[i][1] = t._palette[i].g;
                             paletteView[i][2] = t._palette[i].b;
                         }
-                        int shape[2] = { 256, 3 };
+                        size_t shape[2] = { 256, 3 };
                         return nb::cast(nb::ndarray<nb::numpy, unsigned char, nb::shape<256, 3>>(paletteView, 2, shape)); 
                     },
                     [](apngasm::APNGFrame &t, nb::ndarray<unsigned char, nb::shape<256, 3>> *v) APNGASM_PY_API {
@@ -254,7 +254,7 @@ NB_MODULE(MODULE_NAME, m) {
 
         .def_prop_rw("transparency", 
                     [](apngasm::APNGFrame &t) APNGASM_PY_API {
-                        int shape[1] = { t._transparencySize };
+                        size_t shape[1] = { static_cast<size_t>(t._transparencySize) };
                         return nb::cast(nb::ndarray<nb::numpy, unsigned char, nb::shape<nb::any>>(t._transparency, 1, shape)); 
                     },
                     [](apngasm::APNGFrame &t, nb::ndarray<unsigned char, nb::shape<nb::any>> *v) APNGASM_PY_API {
